Allow float numbers

+++ calc.py
@@ -1,4 +1,4 @@
 import sys
 
 if __name__ == '__main__':
-   print(sum(map(int, sys.argv[1:])))
+   print(sum(map(float, sys.argv[1:])))
